[front/components/members] - refactor: extract EditInvitationModal to its own component
 - Moved EditInvitationModal from InvitationModal.tsx to EditInvitationModal.tsx to separate concerns
 - Updated import paths in InvitationsList.tsx to reflect the relocation of EditInvitationModal

[front/lib] - feature: implement invitation update and resend logic

 - Added updateInvitation function to handle invitation role updates and revocation
 - Created sendInvitations function to manage sending new invitations and resending existing ones with proper error handling

[front/components/members] - refactor: clean up InvitationModal.tsx

 - Removed EditInvitationModal related code due to its extraction into a separate file
 - Cleaned up unused import statements and functions, simplifying the InvitationModal component

[front/lib] - refactor: move MAX_UNCONSUMED_INVITATIONS_PER_WORKSPACE_PER_DAY constant to invitations.ts

 - Centralized the invitation limit constant for better maintainability
